Put exceptions in separate file
